🔨 Refactor duplicate expression extraction methods into a util function
